Move CreateMACBinding to libovsdb/util pkg
In order to remove dependency of util pkg
on libovsdbops pkg, let's move the utilities
that require libovsdbops in their own suite.
