+++ go-controller/pkg/ovn/gateway/gateway.go
@@ -8,6 +8,7 @@ import (
 	libovsdbclient "github.com/ovn-org/libovsdb/client"
 
 	libovsdbops "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdb/ops"
+	libovsdbutil "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdb/util"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nbdb"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/node"
 	ovntypes "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
@@ -68,7 +69,7 @@ func CreateDummyGWMacBindings(sbClient libovsdbclient.Client, nodeName string) e
 		dummyNextHopMAC := util.IPAddrToHWAddr(nextHop)
 		nodeGWRouter := util.GetGatewayRouterFromNode(nodeName)
 		logicalPort := ovntypes.GWRouterToExtSwitchPrefix + nodeGWRouter
-		if err := util.CreateMACBinding(
+		if err := libovsdbutil.CreateMACBinding(
 			sbClient,
 			logicalPort,
 			nodeGWRouter,
